Move old scripts to score subfolder
