--- opal/conf/__init__.py
@@ -1,3 +1,0 @@
-from opal.conf.conf import OSU_DS_2022_04, DATA_DIR
-
-__all__ = ['OSU_DS_2022_04', 'DATA_DIR']
