--- opal/preprocessing_dynamic/__init__.py
@@ -1,3 +1,0 @@
-from opal.preprocessing_dynamic.preprocessing_dynamic import PreprocessingDynamic
-
-__all__ = ['PreprocessingDynamic']
