--- opal/preprocessing_static/__init__.py
@@ -1,8 +1,0 @@
-from opal.preprocessing_static.classify_sv_maps import classify_sv_maps
-from opal.preprocessing_static.filter_scores import filter_scores
-from opal.preprocessing_static.remove_non_mania import remove_non_mania
-from opal.preprocessing_static.filter_beatmaps import filter_beatmaps
-from opal.preprocessing_static.join_score_beatmaps import join_score_beatmaps
-
-__all__ = ['classify_sv_maps', 'filter_scores', 'remove_non_mania',
-           'filter_beatmaps', 'join_score_beatmaps']
