+++ opal/score/conf/__init__.py
@@ -1,0 +1,3 @@
+from opal.score.conf.conf import OSU_DS_2022_04, DATA_DIR
+
+__all__ = ['OSU_DS_2022_04', 'DATA_DIR']
