+++ opal/score/preprocessing_dynamic/__init__.py
@@ -1,0 +1,3 @@
+from opal.score.preprocessing_dynamic.preprocessing_dynamic import PreprocessingDynamic
+
+__all__ = ['PreprocessingDynamic']
